data/macro_data_layer/pyproject.toml
[project]
name = "macro-data-layer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fredapi>=0.5.2",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
